If I want to move, this is what needs to happen. Apartments are starting at about $800 a month near my job. With this plan I can then have an apartment budget between $800 and $970.
Now I can see how it's possible for tears to come to someone's eyes as features are being removed from a project. Depending on the project it could be taking food out of their mouths.

And as you can see, I'm forcing myself to have a perfect budget with zero cash balance, or food will be coming out of my mouth as that's about the only discretionary expense. The petty cash is nothing $80 will satisfy undocumented expenses like toothpaste and full synthetic oil changes once a month.

This will start to matter when projections are built-out. I'm really need $800 or I am going to play it seriously close, but I need to be able to get right up to that line - but not cross it.