master expense config
This is what TDD looks like in practice. I have my entire test with a
huge amount of complexity to build on. This took months to get to
though. I have more work that needs to be finalized that shows why this
isn't complete yet, it's the output. It will go onto a calendar that is
used like a balance sheet for transparency. In order to even have that
in the realm of possibilities I need the input and some expected
outputs. The calendar will cause a complexity explosion! So I'm
approaching it carefully.