weekly loan payoff date calculator, without considering interest.
yearly interest can be turned to monthly interest. then monthly interest can be considered when determining a payoff date.

a monthly interest rate would be "good enough". anything but monthly compounding interest from an APR is going to be complex. is there a standard to say how interest is accrued when making payments more frequently than monthly which has the advertised interest and compound rates. for example a 12% APR, actually says 1% every month. then if I wanted to accurately accrue at a weekly only rate it would have to make principal only payments after going over the minimum, for the current month. that's wildly complex.

even when calculating interest, it's not complete, because the concept of a principal only payment would have to be introduced. this is going to get complicated very very quickly.